+++ apps/webapp/app/components/runs/TimeFrameFilter.tsx
@@ -7,6 +7,7 @@ import { DateField } from "../primitives/DateField";
 import { formatDateTime } from "../primitives/DateTime";
 import { Paragraph } from "../primitives/Paragraph";
 import { Popover, PopoverContent, PopoverTrigger } from "../primitives/Popover";
+import { Label } from "../primitives/Label";
 
 type RunTimeFrameFilterProps = {
   from?: number;
@@ -190,7 +191,7 @@ const timeFrameValues = [
 
 export type RelativeTimeFrameItem = (typeof timeFrameValues)[number];
 
-function AbsoluteTimeFrame({
+export function AbsoluteTimeFrame({
   from,
   to,
   onValueChange,
@@ -202,7 +203,8 @@ function AbsoluteTimeFrame({
   return (
     <div className="flex flex-col gap-2 pt-2">
       <div className="flex flex-col justify-start gap-2">
-        <div className="flex flex-row items-center gap-1">
+        <div className="space-y-1">
+          <Label>From (UTC)</Label>
           <DateField
             label="From (UTC)"
             defaultValue={from}
@@ -214,7 +216,8 @@ function AbsoluteTimeFrame({
             showClearButton
           />
         </div>
-        <div className="flex flex-row items-center gap-1">
+        <div className="space-y-1">
+          <Label>To (UTC)</Label>
           <DateField
             label="To (UTC)"
             defaultValue={to}