+++ apps/webapp/app/presenters/v3/TestTaskPresenter.server.ts
@@ -1,4 +1,10 @@
-import { TaskRunAttemptStatus, TaskRunStatus } from "@trigger.dev/database";
+import { ScheduledTaskPayload, parsePacket, prettyPrintPacket } from "@trigger.dev/core/v3";
+import {
+  RuntimeEnvironmentType,
+  TaskRunAttemptStatus,
+  TaskRunStatus,
+  TaskTriggerSource,
+} from "@trigger.dev/database";
 import { PrismaClient, prisma } from "~/db.server";
 import { getUsername } from "~/utils/username";
 
@@ -8,7 +14,55 @@ type TestTaskOptions = {
   taskFriendlyId: string;
 };
 
-export type TestTask = Awaited<ReturnType<TestTaskPresenter["call"]>>;
+type Task = {
+  id: string;
+  taskIdentifier: string;
+  filePath: string;
+  exportName: string;
+  friendlyId: string;
+  environment: {
+    id: string;
+    type: RuntimeEnvironmentType;
+    userId?: string;
+    userName?: string;
+  };
+};
+
+export type TestTask =
+  | {
+      triggerSource: "STANDARD";
+      task: Task;
+      runs: StandardRun[];
+    }
+  | {
+      triggerSource: "SCHEDULED";
+      task: Task;
+      runs: ScheduledRun[];
+    };
+
+type RawRun = {
+  id: string;
+  number: BigInt;
+  friendlyId: string;
+  createdAt: Date;
+  status: TaskRunStatus;
+  payload: string;
+  payloadType: string;
+  runtimeEnvironmentId: string;
+};
+
+export type StandardRun = Omit<RawRun, "number"> & {
+  number: number;
+};
+
+export type ScheduledRun = Omit<RawRun, "number" | "payload"> & {
+  number: number;
+  payload: {
+    timestamp: Date;
+    lastTimestamp?: Date;
+    externalId?: string;
+  };
+};
 
 export class TestTaskPresenter {
   #prismaClient: PrismaClient;
@@ -17,13 +71,14 @@ export class TestTaskPresenter {
     this.#prismaClient = prismaClient;
   }
 
-  public async call({ userId, projectSlug, taskFriendlyId }: TestTaskOptions) {
+  public async call({ userId, projectSlug, taskFriendlyId }: TestTaskOptions): Promise<TestTask> {
     const task = await this.#prismaClient.backgroundWorkerTask.findFirstOrThrow({
       select: {
         id: true,
         filePath: true,
         exportName: true,
         slug: true,
+        triggerSource: true,
         runtimeEnvironment: {
           select: {
             id: true,
@@ -47,18 +102,7 @@ export class TestTaskPresenter {
       },
     });
 
-    const latestRuns = await this.#prismaClient.$queryRaw<
-      {
-        id: string;
-        number: BigInt;
-        friendlyId: string;
-        createdAt: Date;
-        status: TaskRunStatus;
-        payload: string;
-        payloadType: string;
-        runtimeEnvironmentId: string;
-      }[]
-    >`
+    const latestRuns = await this.#prismaClient.$queryRaw<RawRun[]>`
     WITH taskruns AS (
       SELECT 
           tr.* 
@@ -88,32 +132,63 @@ export class TestTaskPresenter {
     FROM 
         taskruns AS taskr
     WHERE
-        taskr."payloadType" = 'application/json'
+        taskr."payloadType" = 'application/json' OR taskr."payloadType" = 'application/super+json'
     ORDER BY
         taskr."createdAt" DESC;`;
 
-    return {
-      task: {
-        id: task.id,
-        taskIdentifier: task.slug,
-        filePath: task.filePath,
-        exportName: task.exportName,
-        friendlyId: taskFriendlyId,
-        environment: {
-          id: task.runtimeEnvironment.id,
-          type: task.runtimeEnvironment.type,
-          userId: task.runtimeEnvironment.orgMember?.user.id,
-          userName: getUsername(task.runtimeEnvironment.orgMember?.user),
-        },
+    const taskWithEnvironment = {
+      id: task.id,
+      taskIdentifier: task.slug,
+      filePath: task.filePath,
+      exportName: task.exportName,
+      friendlyId: taskFriendlyId,
+      environment: {
+        id: task.runtimeEnvironment.id,
+        type: task.runtimeEnvironment.type,
+        userId: task.runtimeEnvironment.orgMember?.user.id,
+        userName: getUsername(task.runtimeEnvironment.orgMember?.user),
       },
-      runs: latestRuns.map((r) => {
-        //we need to format the code on the server, because we detect if the sample has been edited by comparing the contents
-        try {
-          r.payload = JSON.stringify(JSON.parse(r.payload ?? ""), null, 2);
-        } catch (e) {}
-
-        return { ...r, number: Number(r.number) };
-      }),
     };
+
+    switch (task.triggerSource) {
+      case "STANDARD":
+        return {
+          triggerSource: "STANDARD",
+          task: taskWithEnvironment,
+          runs: await Promise.all(
+            latestRuns.map(async (r) => {
+              const number = Number(r.number);
+
+              return {
+                ...r,
+                number,
+                payload: await prettyPrintPacket(r.payload, r.payloadType),
+              };
+            })
+          ),
+        };
+      case "SCHEDULED":
+        return {
+          triggerSource: "SCHEDULED",
+          task: taskWithEnvironment,
+          runs: await Promise.all(
+            latestRuns.map(async (r) => {
+              const number = Number(r.number);
+
+              return {
+                ...r,
+                number,
+                payload: await getScheduleTaskRunPayload(r),
+              };
+            })
+          ),
+        };
+    }
   }
+}
+
+async function getScheduleTaskRunPayload(run: RawRun) {
+  const payload = await parsePacket({ data: run.payload, dataType: run.payloadType });
+  const parsed = ScheduledTaskPayload.parse(payload);
+  return parsed;
 }