+++ packages/database/prisma/migrations/20240416100646_add_cron_description_column/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "TaskSchedule" ADD COLUMN     "cronDescription" TEXT NOT NULL DEFAULT '';