+++ packages/core-apps/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @trigger.dev/core-apps
 
+## 3.0.0-beta.13
+
 ## 3.0.0-beta.12
 
 ## 3.0.0-beta.11